+++ dataloom/statements/__init__.py
@@ -213,6 +213,7 @@ class GetStatement[T]:
         orders: Optional[list[str]] = [],
         groups: list[tuple[str]] = [],
         having: list[str] = [],
+        distinct: bool = False,
     ):
         (group_columns, group_fns) = groups
         options = [
@@ -228,6 +229,7 @@ class GetStatement[T]:
                 table_name=f'"{self.table_name}"',
                 filters=" ".join(placeholder_filters),
                 options=" ".join(options),
+                distinct="DISTINCT" if distinct else "",
             )
         elif self.dialect == "mysql":
             sql = MySqlStatements.SELECT_WHERE_COMMAND.format(
@@ -235,6 +237,7 @@ class GetStatement[T]:
                 table_name=f"`{self.table_name}`",
                 filters=" ".join(placeholder_filters),
                 options=" ".join(options),
+                distinct="DISTINCT" if distinct else "",
             )
         elif self.dialect == "sqlite":
             sql = Sqlite3Statements.SELECT_WHERE_COMMAND.format(
@@ -242,6 +245,7 @@ class GetStatement[T]:
                 table_name=f"`{self.table_name}`",
                 filters=" ".join(placeholder_filters),
                 options=" ".join(options),
+                distinct="DISTINCT" if distinct else "",
             )
         else:
             raise UnsupportedDialectException(
@@ -258,6 +262,7 @@ class GetStatement[T]:
         orders: Optional[list[str]] = [],
         groups: list[tuple[str]] = [],
         having: list[str] = [],
+        distinct: bool = False,
     ):
         (group_columns, group_fns) = groups
         options = [
@@ -273,18 +278,21 @@ class GetStatement[T]:
                 column_names=", ".join([f'"{name}"' for name in fields] + group_fns),
                 table_name=f'"{self.table_name}"',
                 options=" ".join(options),
+                distinct="DISTINCT" if distinct else "",
             )
         elif self.dialect == "mysql":
             sql = MySqlStatements.SELECT_COMMAND.format(
                 column_names=", ".join([f"`{name}`" for name in fields] + group_fns),
                 table_name=f"`{self.table_name}`",
                 options=" ".join(options),
+                distinct="DISTINCT" if distinct else "",
             )
         elif self.dialect == "sqlite":
             sql = Sqlite3Statements.SELECT_COMMAND.format(
                 column_names=", ".join([f"`{name}`" for name in fields] + group_fns),
                 table_name=f"`{self.table_name}`",
                 options=" ".join(options),
+                distinct="DISTINCT" if distinct else "",
             )
         else:
             raise UnsupportedDialectException(