+++ dataloom/tests/sqlite3/test_query_sqlite.py
@@ -581,3 +581,77 @@ class TestQueryingSQLite:
         assert len(post) == 3
 
         conn.close()
+
+    def test_distinct(self):
+        from dataloom import (
+            Loom,
+            Model,
+            Column,
+            PrimaryKeyColumn,
+            CreatedAtColumn,
+            TableColumn,
+            ForeignKeyColumn,
+            Filter,
+            ColumnValue,
+        )
+
+        sqlite_loom = Loom(dialect="sqlite", database="hi.db")
+
+        class User(Model):
+            __tablename__: TableColumn = TableColumn(name="users")
+            id = PrimaryKeyColumn(type="int", auto_increment=True)
+            name = Column(type="text", nullable=False, default="Bob")
+            username = Column(type="varchar", unique=True, length=255)
+            tokenVersion = Column(type="int", default=0)
+
+        class Post(Model):
+            __tablename__: TableColumn = TableColumn(name="posts")
+            id = PrimaryKeyColumn(
+                type="int", auto_increment=True, nullable=False, unique=True
+            )
+            completed = Column(type="boolean", default=False)
+            title = Column(type="varchar", length=255, nullable=False)
+            # timestamps
+            createdAt = CreatedAtColumn()
+            # relations
+            userId = ForeignKeyColumn(
+                User,
+                maps_to="1-N",
+                type="int",
+                required=True,
+                onDelete="CASCADE",
+                onUpdate="CASCADE",
+            )
+
+        conn, tables = sqlite_loom.connect_and_sync([User, Post], drop=True, force=True)
+        userId = sqlite_loom.insert_one(
+            instance=User,
+            values=ColumnValue(name="username", value="@miller"),
+        )
+        for title in ["Hey", "Hello", "What are you doing", "Coding"]:
+            sqlite_loom.insert_one(
+                instance=Post,
+                values=[
+                    ColumnValue(name="userId", value=userId),
+                    ColumnValue(name="title", value=title),
+                ],
+            )
+
+        post = sqlite_loom.find_many(
+            Post,
+            filters=Filter(
+                column="id",
+                operator="not",
+                value=2,
+            ),
+            select=["completed"],
+            distinct=True,
+        )
+        assert len(post) == 1
+        post = sqlite_loom.find_all(
+            Post,
+            select=["completed"],
+            distinct=True,
+        )
+        assert len(post) == 1
+        conn.close()